Code showing choreographe's bug
The bug: the variable gr1_score, although being updated in ALMemory by the python box BEFORE being read in the proposal - is read as the previous value, even though a trigger of another function in the python box reads the correct value before and
after qiChat reads the wrong one.
To replicate run this commit's programme (you can use the virtual robot)
and type in the following input, line by line when promped to/it is
appropriate:
>Lets Play
>Trivia
>Tigers
>Owls
(wait for the prompt of the question - "What's 1 plus 1")
>Tigers 2
(you will see that "Tigers has 5 points" (the initial value) but in the
logger, gr1_score=6 from both the triggers of $tryVal=0 before and after
the reading of $gr1_score.
$tryVal=0 triggers the function onInput_tryVal which reads gr1_score from
ALMemory in Python - and gets the updated value (6 and not 5) not like the
qiChat's reading)